+++ rwt/scripts.py
@@ -15,6 +15,10 @@ if sys.version_info < (3,):
 	filter = itertools.ifilter
 
 
+class Dependencies(list):
+	index_url = None
+
+
 class DepsReader:
 	"""
 	Given a Python script, read the dependencies it declares.
@@ -39,7 +43,7 @@ class DepsReader:
 			reader = cls.load(script_path)
 			return reader.read()
 		except Exception:
-			return []
+			return Dependencies()
 
 	@classmethod
 	def search(cls, params):
@@ -58,7 +62,8 @@ class DepsReader:
 		['foo']
 		"""
 		reqs_raw = self._read('__requires__')
-		return list(map(str, pkg_resources.parse_requirements(reqs_raw)))
+		strings = map(str, pkg_resources.parse_requirements(reqs_raw))
+		return Dependencies(strings)
 
 	def _read(self, var_name):
 		mod = ast.parse(self.script)