+++ CHANGES.rst
@@ -6,7 +6,9 @@ Minor incompatibilty - ``DepsReader.read`` no longer accepts a
 
 #19: DepsReader.read and DepsReader.try_read now return a
 scripts.Dependencies instance, which always has an
-``index_url`` attribute (may be None).
+``index_url`` attribute whose value will reflect
+the value of ``__index_url__`` from the script (if present)
+or None otherwise.
 
 2.16
 ----